Add attribution for parseRepositoryTag
